start: create structure of the train_split/tokenization project.
Here the MLproject file with 3 entry_points is created:
- install_language_models
- tokenize_data
- split_and_upload
Also, the skelleton of what will be the `split_data/conda.yaml` also
is created as well as the `splitting.py` and `tokenization.py` file.

Finally, all the steps of the pipeline except for the split_data at the
`main.py` file were commented.