fix: correct entrypoints for all the steps
here, the previously not-implemented entrypoints are now implemented.
The only thing that makes it works is that, instead of receiving all
possible overrides, it only receives the fixed one that tells main.py
that the only step to execute from the full pipeline is the one
related to that respective entrypoint.