refactor: redo the saving way of the tf models
Before this modification, the responsible for saving the tf models was
WandbCallback() at the training call. With this, more than only one model
was being saved at W&B, which as problematic since the data storage was being
fastly filled. Now, the model save is made manually (and note that, the format
which it is being save is `SavedModel` format).

https://www.tensorflow.org/guide/keras/save_and_serialize?hl=en